ci: move access control test suite
